refactor(models): reorganize conference and team models, update fields and dependencies
